NEW: adds Artifact cache, pools, and extra-process GC; replaces qiime2-archive-* tmpdirs (#623)
This commit adds new objects `qiime2.Cache`, `qiime2.Pool` which control the placement of QIIME 2 artifacts on disk.

This commit also removes qiime2-archive-* temporary directories as artifacts are now loaded directly into the currently specified cache. There is now a garbage collector available to the cache which will handle cleanup even in the event of ungraceful shutdown of the original process (such as a `kill -9`), for now these orphaned references will be removed 45 days after creation.

+++ qiime2/core/archive/provenance.py
@@ -323,9 +323,13 @@ class ProvenanceCapture:
         # condition on `rename`, so fall back to copying.
         try:
             os.rename(self.path, final_path)
-        except FileExistsError:
-            distutils.dir_util.copy_tree(str(self.path), str(final_path))
-            distutils.dir_util.remove_tree(str(self.path))
+        except (FileExistsError, OSError) as err:
+            if isinstance(err, FileExistsError) or isinstance(err, OSError) \
+                 and err.errno == 18:
+                distutils.dir_util.copy_tree(str(self.path), str(final_path))
+                distutils.dir_util.remove_tree(str(self.path))
+            else:
+                raise err
 
     def fork(self):
         forked = copy.copy(self)
